backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
httpx==0.27.0
openai==1.40.0
reportlab==4.2.2
pgvector==0.2.5
alembic==1.13.2
tenacity==8.5.0
python-multipart==0.0.9
loguru==0.7.2